[file] [rrr-socket-client] Soft error was ignored in socket client collection
This happened also when socket surely was closed, like when 0 bytes were read.
This caused the socket not to be closed with following CPU spin due do constant POLLIN event.
The soft error is now forced to propagate if bytes read is zero so that only data errors are
ignored (which is also intention of higher levels).