[js] [cmodule] [stats] Implement stats messages from forks
Workers normally cannot communicate with stats engine, and if they post
messages to the stats engine, the messages gets stuck in the memory of
the engine (as it does not run in the fork).

Implement passing of stats messages from worker forks by hooking them in
the stats instance framework and sending them on the mmap channel.

Currently, only the JS module does this. To support more modules, the
stats instance framework should probably have functionallity moved out
as the worker shuold not need to use the stats instance to make messages
as this object is not usable outside of the main process.